Move test cases to internal.
