+++ internal/test/t1/lexer/acttab.go
@@ -5,7 +5,7 @@ package lexer
 import (
 	"fmt"
 
-	"github.com/goccmack/gocc/test/t1/token"
+	"github.com/goccmack/gocc/internal/test/t1/token"
 )
 
 type ActionTable [NumStates]ActionRow
