+++ internal/test/t1/lexer/transitiontable.go
