+++ internal/test/t1/parser/parser.go
@@ -6,8 +6,8 @@ import (
 	"bytes"
 	"fmt"
 
-	parseError "github.com/goccmack/gocc/test/t1/errors"
-	"github.com/goccmack/gocc/test/t1/token"
+	parseError "github.com/goccmack/gocc/internal/test/t1/errors"
+	"github.com/goccmack/gocc/internal/test/t1/token"
 )
 
 const (
