+++ internal/test/t1/terminals.txt
