+++ internal/test/t1/util/litconv.go
