--- test/Makefile
@@ -1,2 +1,0 @@
-regenerate:
-	(cd t1 && make regenerate)
